+++ packages/ui/CHANGELOG.md
@@ -5,7 +5,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-## [0.27.0] - 2026-09-22
+## [0.27.0]
 
 ### Changed